refactor(harness): rename compatibility → prometheus-compliance
Makes room for sibling harness/loki-compliance + harness/tempo-compliance
dirs per docs/loki-compliance-plan.md and docs/tempo-compliance-plan.md
(merged in #332).

- git mv harness/compatibility harness/prometheus-compliance
  (preserves history)
- update all path references: Justfile, docs/*.md, CLAUDE.md,
  CONTRIBUTING.md, CHANGELOG.md, .github/workflows/*.yml,
  .markdownlintignore, .gitmodules, cmd/cerberus/main_test.go,
  test/property/doc.go, harness/prometheus-compliance/scripts/*

The .github/workflows/compatibility.yml filename + the job name
"prometheus/compliance" are unchanged — branch protection's
required-status-checks list pins the latter.
