+++ CLAUDE.md
@@ -9,7 +9,7 @@ Drop-in **Prometheus / Loki / Tempo** HTTP gateway for **ClickHouse**. Parses ea
 - **Conventional Commits**, enforced by `commitlint` (see `.commitlintrc.json`). The `subject-case` rule is relaxed so Dependabot's `Bump X from Y to Z` subjects pass.
 - **Justfile is the canonical task runner.** `just` lists every recipe. Don't reach for `go test ./...` directly when `just test` exists — the recipe sets the race flag, the cover profile, and the right toolchain.
 - **No local validation; lefthook + CI own it.** Don't run `just test`, `just lint`, `go test`, `golangci-lint run`, `go build`, or `markdownlint-cli2` manually as a pre-flight before pushing. The repo's `lefthook.yml` runs lightweight formatters (`gofumpt` / `goimports` / `markdownlint-cli2 --fix`) on staged files at commit time, and the `commit-msg` hook validates Conventional Commits via `commitlint`. CI (`check` + `lint` jobs) runs the heavy validation. New contributors run `just hooks-install` once after cloning; agents trust the hook + CI and don't pre-flight.
-- **Compatibility is the source of truth for PromQL.** `compatibility.yml` runs on main pushes + nightly + manual dispatch and acts as the informational baseline; a future cut can re-enable the `pull_request:` trigger and add `prometheus/compliance` to required checks. An entry in `harness/compatibility/expected-failures.json` requires a comment explaining the upstream rationale.
+- **Compatibility is the source of truth for PromQL.** `compatibility.yml` runs on main pushes + nightly + manual dispatch and acts as the informational baseline; a future cut can re-enable the `pull_request:` trigger and add `prometheus/compliance` to required checks. An entry in `harness/prometheus-compliance/expected-failures.json` requires a comment explaining the upstream rationale.
 - **No raw SQL strings — typed chsql API only.** Use `internal/chsql.Builder` / `chsql.QueryBuilder` — the custom CH-flavored builder API (see [`docs/sql-builder-evaluation.md`](docs/sql-builder-evaluation.md) for the build-vs-buy decision and [`docs/roadmap.md` § RC6](docs/roadmap.md#rc6--type-safe-sql-via-custom-internalchsqlbuilder) for the milestone sequence). Compose clauses via typed `QueryBuilder` slots (`.Select` / `.From` / `.Where` / `.GroupBy` / `.OrderBy` / `.Limit` / `.Prewhere` / `.Join` / `.WithRecursive`) and expressions via typed Frags (`Eq` / `And` / `Or` / `Paren` / `Cast` / `In` / `Like` / `Add` / `Call` / `Array` / `Subscript` / `If` / `Lambda1` / `Subquery` / `BareIdent` / `InlineLit` / etc.). `Builder.writeSQL` is unexported and the `chsql.Raw` / `chsql.Concat` public escape hatches are gone — external packages cannot raw-write SQL by construction. The typed-Frag surface is closed: add new typed constructors when a shape isn't covered, never compose SQL via string concatenation. Reviewer discipline + the typed API are the enforcement.
 
 ## Architecture map
@@ -30,7 +30,7 @@ test/property/               oracle-based property tests (`pgregory.net/rapid` s
 test/regression/             meta-tests that pin past CI failures so they can't silently recur — goleak detectors across every handler entrypoint (added by #253), justfile-shape pins, seed-program invariants.
 test/e2e/                    k3d cluster + Grafana playwright smoke
 test/e2e/{k3s,grafana}/      k3d manifests + Grafana provisioning (datasources, dashboards) consumed by the smoke
-harness/compatibility/          prometheus/compliance Docker Compose harness + shadow-mode differential testing
+harness/prometheus-compliance/          prometheus/compliance Docker Compose harness + shadow-mode differential testing
 docs/                        roadmap.md, optimizer-research.md, compatibility.md, engine.md, observability.md, 12factor.md, test-strategy.md, …
 ```
