+++ harness/prometheus-compliance/shadow/corpus/smoke.txt
