+++ harness/prometheus-compliance/shadow/differ.go
