+++ harness/prometheus-compliance/shadow/result_adapter.go
